Create basic discord client
At this point, we just connect and print out that we're logged in when
we're ready. There's no behavior right now for actually interacting with
Discord; that will come later.

Note that, for privacy reasons, the secret has not been committed and
should be passed in via the first argument or as an environment variable
`DISCORD_TOKEN`.