Add simple database for users
Just as a test to make sure I can store users with their guild, I've
created the wrapper class for the SQL database (as `pearbot.database`)
and added two methods : `register_user` (to register a user) and `users`
(to get all user data in the table). The `users` method may be too
permissive, just returning a cursor, but for now it seems better to
return the cursor than to put it in a list, so that the caller can
iterate through the values at their own pace.

The `database.Connection` class forwards its `close` to the actual
connection object, making it valid to put in a `closing` context
manager. The `database.TransactionManager` manages the database
transactions, which are normally handled merely by the connection as a
context manager. Users cannot perform actions on the database without
starting a transaction, making for safer behavior.