Swap direct connection calls to cursor calls
The `sqlite3` library has some convenience methods on its connection
object so that you can skip getting a cursor explicitly. However, this
is not a standard SQL method, so if we switched to a different SQL
implementation later on, it would break. This change ensures that it
will work with all compliant SQL implementations.